feat: separate TextField and InputField
